requirements.txt
colorlog==6.9.0
fastapi==0.115.8
uvicorn==0.34.0
pymongo==4.11
dnspython==2.7.0
pydantic==2.9.2
pydantic[email]==2.9.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
bcrypt==4.2.1
pyjwt==2.10.1
pillow==11.0.0
requests==2.32.3
pytest==8.3.4
pytest-ordering==0.6
httpx==0.27.2
pytest-asyncio==0.25.3
pytest-mock==3.14.0
starlette==0.45.3
